+++ src/@types/index.ts
@@ -1,0 +1,15 @@
+export enum Network {
+  MAINNET = 'MAINNET',
+  GOERLI = 'GOERLI',
+}
+
+export type StartParams = {
+  readonly network?: Network;
+  readonly rpc_port?: number;
+  readonly untrusted_rpc_url: string;
+  readonly consensus_rpc_url: string;
+};
+
+export type StartResult = {
+  readonly shutdown: () => Promise<void>;
+};
